// src/types/types.ts
export interface State {
  output: string;
  firstNumber: string;
  secondNumber: string;
  operation: string;
  equalOperation: string;
  status: boolean;
  isEqualed: boolean;
  clearSecondNumber: boolean;
  canDelete: boolean;
  disabled: boolean;
}

// export interface Colors {
//   "--background-color": string;
//   "--background-result-color": string;
//   "--background-buttons-color": string;
//   "--numbers-button-background-color": string;
//   "--numbers-button-text-color": string;
//   "--numbers-button-border-color": string;
//   "--equal-button-background-color": string;
//   "--equal-button-text-color": string;
//   "--equal-button-border-color": string;
//   "--rest-button-background-color": string;
//   "--rest-button-text-color": string;
//   "--rest-button-border-color": string;
//   "--text-color": string;
// }

// export interface ThemeColors {
//   darkmode: Colors;
//   lightmode: Colors;
// }

// export type Theme = "darkmode" | "lightmode";